+++ neuronquant/data/ziplinesources/live/forex.py
@@ -18,6 +18,7 @@
 Tools to generate data sources.
 """
 import time
+import datetime
 import pandas as pd
 
 from zipline.gens.utils import hash_args
@@ -26,7 +27,7 @@ from zipline.sources.data_source import DataSource
 from neuronquant.data.forex import ConnectTrueFX
 
 import logbook
-log = logbook.Logger('DataLiveSource')
+log = logbook.Logger('ForexLiveSource')
 
 
 class DataLiveSource(DataSource):
@@ -53,12 +54,13 @@ class DataLiveSource(DataSource):
         self.start = kwargs.get('start', data['index'][0])
         self.end   = kwargs.get('end', data['index'][-1])
 
-        self.fake_index = pd.date_range(self.start, self.end, freq=pd.datetools.Minute())
+        #self.fake_index = pd.date_range(self.start, self.end, freq=pd.datetools.Minute())
 
         # Hash_value for downstream sorting.
         self.arg_string = hash_args(data, **kwargs)
 
-        self._raw_data = None
+        self._raw_data = kwargs.pop('data_gen') if 'data_gen' in kwargs else None
+        #self._raw_data = None
 
         assert isinstance(self.sids, list)
         self.forex = ConnectTrueFX(pairs=self.sids)
@@ -78,12 +80,36 @@ class DataLiveSource(DataSource):
             'volume': (int, 'volume')
         }
 
+    def _wait_for_dt(self, dt):
+        '''
+        Only return when we reach given datetime
+        '''
+        current_dt = datetime.datetime.now()
+        while (current_dt.minute < dt.minute) or (current_dt.hour < dt.hour) :
+            time.sleep(15)
+            current_dt = datetime.datetime.now()
+            log.info('Waiting {} / {}'.format(current_dt, dt))
+
+    def _get_updated_index(self):
+        '''
+        truncate past dates in index
+        '''
+        late_index = self.data['index']
+        current_dt = datetime.datetime.now()
+        selector = (late_index.day > current_dt.day) \
+                | ((late_index.day == current_dt.day) & (late_index.hour > current_dt.hour)) \
+                | ((late_index.day == current_dt.day) & (late_index.hour == current_dt.hour) & (late_index.minute >= current_dt.minute))
+        return self.data['index'][selector]
+
     @property
     def instance_hash(self):
         return self.arg_string
 
     def raw_data_gen(self):
-        for fake_dt in self.fake_index:
+        index = self._get_updated_index()
+        #for fake_dt in self.fake_index:
+        for dt in index:
+            self._wait_for_dt(dt)
             while True:
                 log.debug('Waiting for Forex update')
                 currencies = self.forex.QueryTrueFX()
@@ -91,14 +117,14 @@ class DataLiveSource(DataSource):
                 if len(currencies.keys()) >= len(self.sids):
                     log.debug('New income data, fire an event !')
                     break
-                time.sleep(60)
+                time.sleep(30)
 
             try:
                 for sid in self.sids:
                     assert sid in currencies.columns
                     log.debug('Data available:\n{}'.format(currencies[sid]))
                     event = {
-                        'dt': fake_dt,
+                        'dt': dt,
                         'trade_time': currencies[sid]['TimeStamp'],
                         'sid': sid,
                         'bid': currencies[sid]['Bid.Price'],
@@ -106,7 +132,7 @@ class DataLiveSource(DataSource):
                         'high': currencies[sid]['High'],
                         'low': currencies[sid]['Low'],
                         #'price': currencies[sid]['Bid.Price'],
-                        'volume': 0
+                        'volume': 10000
                     }
                     yield event
             except: