+++ server/client_console.js
@@ -14,6 +14,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+
+var zmq = require('zmq')
+    , ui = require('console_ui')
+    , program = require('commander')
+    , consign = require('commander')
+    , config = require('config')
+    , broker_uri = config.network.frontport
+    , voice = require('vocal');
+
+program
+    .version('0.0.1')
+    .usage('[commands] <args>')
+    .description('Remote client for NeuronQuant trading system')
+    .option('-v, --verbose <level>', 'verbosity', Number, 2)
+    .option('-c, --channel <name>', 'Channel name of the client', String, 'dashboard')
+    .parse(process.argv);
+
+
+//TODO Move it to default.json, read by config ?
 var bt_config = {
     type: "fork",
     script: "backtester/backtest.py",
@@ -21,11 +40,11 @@ var bt_config = {
     args: {
         cash: {
             prefix: "-i",
-            value: 100000 
+            value: 10000 
         } ,
         ticker: {
             prefix: "--ticker",
-            value: "google,apple" 
+            value: "random,2" 
         },
         exchange: {
             prefix: "--exchange",
@@ -40,8 +59,8 @@ var bt_config = {
             value: "BuyAndHold" 
         },
         delta: {
-            prefix: "--delta",
-            value: '1min'
+            prefix: "--frequency",
+            value: 'minute'
         },
         manager: {
             prefix: "--manager",
@@ -53,11 +72,15 @@ var bt_config = {
         },
         end: {
             prefix: "--end",
-            value: "2008-07-03" 
+            value: "19h20" 
         } ,
-        mode: {
+        client: {
             prefix: "flag",
             value: "--remote"
+        },
+        mode: {
+            prefix: "flag",
+            value: "--live"
         } 
     },
     configuration: {
@@ -68,11 +91,12 @@ var bt_config = {
             threshold: 0
         },
         manager: {
-            load_backup: 0,
+            name: "Xavier Bruhiere",
+            load_backup: 1,
             max_weight: 0.5,
             connected: 1,
-            buy_amount: 200, 
-            sell_amount: 100
+            buy_amount: 80, 
+            sell_amount: 70
         }
     },
     monitor: "notImplemented",
@@ -109,28 +133,16 @@ var opt_config = {
     }
 }
 
-
-var zmq = require('zmq')
-    , ui = require('client_ui')
-    , program = require('commander')
-    , config = require('config')
-    , broker_uri = config.network.frontport
-    , voice = require('vocal');
-
-program
-    .version('0.0.1')
-    .usage('[commands] <args>')
-    .description('Remote client for NeuronQuant trading system')
-    .option('-v, --verbose <level>', 'verbosity', Number, 2)
-    .option('-c, --channel <name>', 'Channel name of the client', String, 'dashboard')
-    .parse(process.argv);
-
 var fd_config = {
     type: "configure",
     filters: ["portfolio", "acknowledgment", "optimization", "vocal"],
     log_redirection: program.channel,
     verbose: program.verbose
 };
+
+var mega_config = {'backtest': bt_config,
+                   'broker': fd_config,
+                   'optimization': opt_config};
     
 /*
  * ZMQ based client, meant to communicate with the server forwarder 
@@ -138,8 +150,8 @@ var fd_config = {
  */
 function create_client (port, channel) {
     //NOTE Could set a flag to make it talk
-    client_ui.write_log('Log window setup done.');
-    client_ui.write_msg('Message window setup done.');
+    console_ui.write_log('Log window setup done.');
+    console_ui.write_msg('Message window setup done.');
 
     var socket = zmq.socket('dealer');
     
@@ -151,26 +163,27 @@ function create_client (port, channel) {
         json_data = JSON.parse(data);
 
         if (json_data.type == 'portfolio') {
-            client_ui.write_msg(json_data.time + ' Portfolio:');
-            client_ui.write_msg(JSON.stringify(json_data.msg.portfolio_value));
+            console_ui.write_msg(json_data.time + ' Portfolio:');
+            //console_ui.write_msg(JSON.stringify(json_data.msg.portfolio_value));
+            console_ui.write_msg(JSON.stringify(json_data.msg));
             
             //FIXME Work only in above line configuration
-            //client_ui.write_msg(json_data.time + ' Returns:', json_data.msg['returns']);
-            //client_ui.write_msg(json_data.time + ' PNL:', json_data.msg.pnl);
+            //console_ui.write_msg(json_data.time + ' Returns:', json_data.msg['returns']);
+            //console_ui.write_msg(json_data.time + ' PNL:', json_data.msg.pnl);
         }
 
         else if (json_data.type == 'optimization') {
-            client_ui.write_msg(json_data.msg['iteration'] + ' ' + json_data.msg['progress'] + '% | ' + json_data.msg['best'] + ' (' + json_data.msg['mean'] +')');
+            console_ui.write_msg(json_data.msg['iteration'] + ' ' + json_data.msg['progress'] + '% | ' + json_data.msg['best'] + ' (' + json_data.msg['mean'] +')');
         }
 
         else if (json_data.type == 'acknowledgment') {
-            client_ui.write_msg(json_data.time + ' - Worker returned: ' + json_data.msg);
+            console_ui.write_msg(json_data.time + ' - Worker returned: ' + json_data.msg);
             socket.send(JSON.stringify({type: 'acknowledgment', statut: 0}));
         }
 
         //TODO Could be a vocal optimization message -> flag in the message or multiple type possible
         else if (json_data.type == 'vocal') {
-            client_ui.write_log(json_data.time + ' ' + json_data.func_name + ': ' + json_data.msg);
+            console_ui.write_log(json_data.time + ' ' + json_data.func_name + ': ' + json_data.msg);
             voice.synthetize(json_data.msg, config.vocal.lang);
         }
 
@@ -178,14 +191,14 @@ function create_client (port, channel) {
             //NOTE Should be dedicated to unexpected message, for now it's for logging
             //NOTE could configure a channel_log_filter in addition to the type filter, or merge them
             //TODO msg can be json object, detect and parse it
-            client_ui.write_log(json_data.time + ' ' + json_data.func_name + ': ' + json_data.msg);
+            console_ui.write_log(json_data.time + ' ' + json_data.func_name + ': ' + json_data.msg);
         }
     });
 
     socket.connect(port);
-    client_ui.write_msg('[Node:Client] ' + socket.identity + ' connected');
+    console_ui.write_msg('[Node:Client] ' + socket.identity + ' connected');
     
-    client_ui.write_msg('Initialize the forwarder with default configuration');
+    console_ui.write_msg('Initialize the forwarder with default configuration');
     socket.send(JSON.stringify(fd_config));
 
     this.send_json = function(msg) {
@@ -199,20 +212,69 @@ function create_client (port, channel) {
 
 // Define client interfaces, ui for curses graphics, be for backend work
 // The test function is a callback called when hitting keyboard
-var client_ui = new ui.UI_interface(test),
+var console_ui = new ui.UI_interface(command),
     client_be = new create_client(broker_uri, program.channel);
 
-function test(msg) {
-    if (msg == 'configure') {
-        client_be.send_json(fd_config);
+
+function isJsonString(text) {
+
+    try {
+        res = JSON.parse(text);
+    } catch (e) {
+        return false;
     }
-    else if (msg == 'backtest') {
-        client_be.send_json(bt_config);
+    return true;
+}
+
+function command(msg) {
+    // We're building a command line alike array from msg,
+    // i.e. 'node program --flag value ...'
+    command_line = msg.split(' ');
+    command_line.unshift('client_console.js');
+    //FIXME A bad argument crashs the program (and screws up the terminal)
+    //FIXME -c doesn't work
+    consign
+        .option('-p, --property <value>', 'json string, to edit specified json config name accordingly', String, '{}')
+        .option('-c, --config <name>', 'Target configuration object', String, 'backtest')
+        .parse(command_line);
+
+    if (command_line[1] == 'send') {
+        if (consign.config in mega_config) {
+            console_ui.write_msg(JSON.stringify(mega_config[consign.config], null, 4))
+            client_be.send_json(mega_config[consign.config]);
+        }
+        else {
+            console_ui.write_msg('** Error: ' + consign.config + ' configuration does not exist');
+        }
     }
-    else if (msg == 'optimize') {
-        client_be.send_json(opt_config);
+
+    else if (command_line[1] == 'order') {
+        if (isJsonString(consign.property)) {
+            console_ui.write_msg('Valid order, sending: ' + consign.property);
+            client_be.send_json(consign.property);
+        } else {
+            console_ui.write_msg('Invalid json string, just pass');
+        }
     }
-    else if (msg == 'order') {
-        client_be.send_json({type: 'test', msg: msg});
+
+    else if (command_line[1] == 'show') {
+        console_ui.write_msg(JSON.stringify(mega_config[consign.config]), null, 4);
+    }
+
+    else if (command_line[1] == 'edit') {
+        //TODO handle depth 2 at least
+        bt_config.monitor = 'pouet';
+
+        if (isJsonString(consign.property)) {
+            new_values = JSON.parse(consign.property);
+            console_ui.write_msg('Whole: ' + JSON.stringify(new_values))
+            for (var key in new_values) {
+                console_ui.write_msg('Value: ' + new_values[key])
+                mega_config[consign.config][key] = new_values[key]
+            }
+
+            console_ui.write_msg('Updating configuration: ');
+            console_ui.write_msg(JSON.stringify(mega_config[consign.config], null, 4));
+        }
     }
 }